StudentScorePrediction/version_1/requirements.txt
scikit-learn==1.2.2
numpy==1.23.5
pandas==2.2.2
scipy==1.12.0
xgboost==2.1.2
pyyaml==6.0.2
